# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Jobsy API
  version: "1.0.0"
  description: API для регистрации, входа, подтверждения e-mail, восстановления пароля и управления профилями фрилансера и заказчика.
...
paths:
  /auth/me:
    get:
      tags:
        - auth
      summary: Получить информацию о текущем пользователе
      description: Возвращает данные пользователя, если запрос содержит действительный JWT-токен. Если токен отсутствует или недействителен, возвращается ошибка 401
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о пользователе получена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Пользователь не аутентифицирован
        '500':
          description: Внутренняя ошибка сервера
...
  /auth/login:
    post:
      tags:
        - auth
      summary: Вход пользователя
      description: Аутентифицирует пользователя и возвращает JWT токен
      requestBody:
        description: Учетные данные для входа
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
      responses:
        '200':
          description: Вход выполнен успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '401':
          description: Неверные учетные данные
        '500':
          description: Внутренняя ошибка сервера
...
  /auth/logout:
    post:
      tags:
        - auth
      summary: Выход пользователя
      description: Завершает сессию пользователя. Если реализован механизм аннулирования JWT, сервер также может добавить токен в черный список.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Пользователь успешно вышел из системы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
...
  /admin/auth/login:
    post:
      tags:
        - admin
        - auth
      summary: Запрос кода для входа админа
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: Код для входа успешно отправлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
        '400':
          description: Неверные параметры запроса
        '404':
          description: Администратор не найден
...
  /admin/auth/refresh:
    post:
      tags:
        - admin
        - auth
      summary: Обновление токена
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: Токен успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '500':
          description: Внутренняя ошибка сервера
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    AdminLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "admin@example.com"
      required:
        - email
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: "user@example.com"
        first_name:
          type: string
          example: "Иван"
        last_name:
          type: string
          example: "Иванов"
        date_birth:
          type: string
          format: date
          example: "2000-01-01"
        phone:
          type: string
          example: "79991234567"
        role:
          type: string
          enum: [client, freelancer, admin]
          example: "client"
        is_verified:
          type: boolean
          example: true
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
...
    AuthenticationRequest:
      type: object
      properties:
        email:
          type: string
          example: "user@example.com"
        password:
          type: string
          example: "SecretPassword123!"
      required:
        - email
        - password
    AuthenticationResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
...
    TokenRefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
      required:
        - refreshToken
    TokenRefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
        refreshToken:
          type: string
          example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
        refreshTokenExpiry:
          type: string
          example: "2025-06-01T12:00:00Z"
    DefaultResponse:
      type: object
      properties:
        message:
          type: string
          example: "Операция выполнена успешно"
...